third_party/blink/web_tests/platform/win/external/wpt/html/canvas/element/text/2d.text.drawing.style.letterSpacing.measure-expected.txt

This is a testharness.js-based test.
FAIL Testing letter spacing with different length units assert_approx_equals: letter spacing doesn't work. expected 110 +/- 0.1 but got 55
Harness: the test ran to completion.
